requirements.txt
flask
pillow
requests
numpy       
scipy
